requirements.txt
requests>=2.31.0
solana>=0.30.2
solders>=0.18.0
anchorpy>=0.18.0
base58>=2.1.1
python-dotenv>=1.0.0
